Refactor x/oracle proto params
The original protobuf params field contained a
map of ClaimParams. Maps in Go are unsorted and
unordered. In addition, json.Marshal does not maintain
order for maps. Hashing the map will return a
different hash depending on the ordering, which is a
source of non-determinism.

A workaround for this is using an ordered map
implementation, which isn't necessary for Enci,
as the new params structure should work fine.